requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alive-progress==3.0.1
alpaca-py==0.8.1
altair==5.0.1
anyio==3.6.2
appdirs==1.4.4
art==5.9
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autotrader==0.12.0
aws-cdk-lib==2.110.1
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.110.1a0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.6.2
bokeh==2.3.2
cachetools==5.3.1
cattrs==23.1.2
certifi==2022.12.7
cffi==1.16.0
cftime==1.6.2
chardet==5.2.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.4.0
chromedriver-binary==119.0.6045.105.0
click==8.1.3
colorama==0.4.6
constructs==10.3.0
contourpy==1.1.0
cryptography==42.0.2
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.9
decorator==5.1.1
defusedxml==0.7.1
dicom2nifti==2.4.9
dirsearch==0.4.3.post1
distlib==0.3.7
docker==7.0.0
docker-pycreds==0.4.0
docopt==0.6.2
extra-streamlit-components==0.1.60
fais==0.0.18
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
finta==1.3
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.9.0
gast==0.4.0
gdown==5.0.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
grapheme==0.6.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
highdicom==0.21.1
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.31.1
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.23
Jinja2==3.1.2
joblib==1.3.1
jsii==1.92.0
jsonschema==4.17.3
keras==2.15.0
kiwisolver==1.4.4
langchain==0.0.221
langchainplus-sdk==0.0.19
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
log-symbols==0.0.14
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
monai==1.2.0
monai-deploy-app-sdk==0.5.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
netCDF4==1.6.4
networkx==3.1
nibabel==5.1.0
nltk==3.8.1
ntlm-auth==1.5.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
Pillow==9.5.0
pillow-jpls==1.2.0
pipenv==2023.7.11
pipreqs==0.4.11
platformdirs==3.9.1
plotly==5.15.0
prometheus-client==0.16.0
protobuf==4.23.3
psutil==5.9.5
publication==0.0.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pydicom==2.3.1
pydot==1.4.2
Pygments==2.15.1
pyhumps==3.8.0
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==24.0.0
pyparsing==3.1.0
PyPDF2==3.0.1
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gdcm==3.0.22
python-telegram-bot==20.3
pytorch-lightning==2.0.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML @ file:///private/tmp/pyyaml-20230718-6387-cxgeib/PyYAML-6.0.1
readchar==4.0.5
regex==2023.3.23
requests==2.28.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
runs==1.2.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.30.0
setproctitle==1.3.2
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spinners==0.0.24
SQLAlchemy==2.0.17
sseclient-py==1.7.2
streamlit==1.24.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-cpu==2.13.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.18
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchmetrics==1.1.1
tornado==6.3.2
tqdm==4.65.0
transformers==4.27.4
tweepy==4.14.0
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.14
validators==0.20.0
virtualenv==20.24.0
virtualenv-clone==0.5.7
wandb==0.15.9
wcwidth==0.2.13
websockets==10.4
Werkzeug==2.3.6
wrapt==1.14.1
xmod==1.8.1
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
